Change into a minor mode.
Idea being, less wheel-reinventing, more complementing what Geiser,
Quack, and Scheme modes offer. Note that in this commit it presumes
Geiser in a couple places.

However, I'm still not in love with the enter! evaluation model -- I
find it fiddly and unpredictable. I still prefer the "DrRacket-style"
run that I implemented in my major mode sandbox. But maybe a better way
to address that is with a pull request to Geiser to add such a mode.